fix: route name and extend class
