+++ libs/options-api/src/test/java/gov/cdc/nbs/option/person/names/list/PersonNamesListRequester.java
@@ -17,6 +17,6 @@ class PersonNamesListRequester {
   }
 
   ResultActions request() throws Exception {
-    return mvc.perform(get("/nbs/api/options/person/names")).andDo(print());
+    return mvc.perform(get("/nbs/api/options/person/stdHivWorker/names")).andDo(print());
   }
 }
